requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0
scikit-learn>=0.20.3
opencv-python>=4.0.0.21
bottle>=0.12.16